+++ src/hashes/md5.rs
@@ -1,3 +1,8 @@
 pub fn crack(word: &str) -> String {
     format!("{:x}", md5::compute(word))
 }
+
+pub fn crack_with_salt(word: &str, salt: &str) -> String {
+    let salted = format!("{}{}", salt, word);
+    format!("{:x}", md5::compute(salted))
+}